replace serialization, deserialization methods in class filework and add docstring
