+++ lab_3/asymmetric/public.pem
@@ -1,9 +1,9 @@
 -----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAt4zJKSrhskgOpyBLWB+c
-vrawVg+Pyqb8d0oW7Uer5JyRT8iuwXZu2QO0VDFzvLjWy6C53HU3fPSQOMO+4w0D
-Yq0ccOJ5A6RbqRfy02A85E81FKVegsCIjuqXDWFz0bYPF0qOkXvLefXLcMZaX3xQ
-JB2fDAaOO7XVBQcctjpRkDI/uER/YDF+9KksNAWYxMXPS+XXe8r2LNduQeu3l5sy
-wLFlTsQs2/yTUOP9N5ctisZwxlqcvPUcoDeRdUAWoZwCga5JzAmusMzDA56lK7fX
-R+9ztQl9lyIdP6cx5HoZezx2GBGJIcV6Gdje2PA8lF7vpmMhhNamto1dZbVGirlu
-mQIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAuurBTRa2izxTHEBseivZ
+0dfel3UgNySTVThZM/4FNDF2Uzb4rj/0i4mbG2fJSQc6ZUax0D3JPrLYGTnumiPf
+FJHn3gI0ZKgrwL5WEGe0351rw0BvTfx1pKVGoknkuR+KwCWPaeE5ClNRLe4vRfm9
+ztxze1RLxtZGIElHxf4kNsO3BrFHbX2dyEzlkxWkzgpV9QS/Q+ztLp6ZZ2k9Phnt
+KVwYq/i1COsAsDWfRTIialnlQjHaiLth0LNfv7kU+S8F+l5/sYLyx0IQSj1GZLEL
+DzW3yy/IApglK48T7TIKmfmniL0CTu0bplxPEGG/7LxM0fysiKf9jGmTGdzwv9CY
+BQIDAQAB
 -----END PUBLIC KEY-----
