+++ lab_3/asymmetric_algorithm.py
@@ -1,7 +1,6 @@
 import logging
 
 from cryptography.hazmat.primitives.asymmetric import rsa, padding
-from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives import hashes
 
 logging.basicConfig(level=logging.INFO)
@@ -39,57 +38,6 @@ class AsymmetricAlgorithm:
         private_key = rsa.generate_private_key(public_exponent=65537, key_size=key_size)
         public_key = private_key.public_key()
         return private_key, public_key
-
-    def serialize_private_key(self, private_key: rsa.RSAPrivateKey) -> None:
-        """
-        Serialize the private key and save it to a file.
-
-        :param private_key: The private key.
-        """
-        try:
-            with open(self.private_key_path, 'wb') as key_file:
-                key_file.write(private_key.private_bytes(encoding=serialization.Encoding.PEM,
-                                                         format=serialization.PrivateFormat.TraditionalOpenSSL,
-                                                         encryption_algorithm=serialization.NoEncryption()))
-        except Exception as e:
-            logging.error(f"Error in serializing private key - {e}")
-    
-    def serialize_public_key(self, public_key: rsa.RSAPublicKey) -> None:
-        """
-        Serialize the public key and save it to a file.
-
-        :param public_key: The public key.
-        """
-        try:
-            with open(self.public_key_path, 'wb') as key_file:
-                key_file.write(public_key.public_bytes(encoding=serialization.Encoding.PEM,
-                                                       format=serialization.PublicFormat.SubjectPublicKeyInfo))
-        except Exception as e:
-            logging.error(f"Error in serializing public key - {e}")
-    
-    def deserialize_private_key(self) -> rsa.RSAPrivateKey:
-        """
-        Deserialize the private key and return it.
-        
-        :return: The deserialized private key.
-        """
-        try:
-            with open(self.private_key_path, 'rb') as key_file:
-                return serialization.load_pem_private_key(key_file.read(), password=None)
-        except Exception as e:
-            logging.error(f"Error in deserializing private key - {e}")
-
-    def deserialize_public_key(self) -> rsa.RSAPublicKey:
-        """
-        Deserialize the public key and return it.
-        
-        :return: The deserialized public key.
-        """
-        try:
-            with open(self.public_key_path, 'rb') as key_file:
-                return serialization.load_pem_public_key(key_file.read())
-        except Exception as e:
-            logging.error(f"Error in serializing public key - {e}")
 
     @staticmethod
     def encrypt_with_public_key(public_key: rsa.RSAPublicKey, text: bytes) -> bytes:
