+++ lab_3/file_work.py
@@ -1,10 +1,20 @@
 import json
 import logging
 
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric import rsa
+
 logging.basicConfig(level=logging.INFO)
 
 
 class FileWork:
+    """
+    A class containing additional functions for reading and writing files, 
+    serializing and deserializing keys.
+
+    Attributes:
+        None
+    """
     def __init__(self, path: str) -> None:
         self.path = path
 
@@ -67,4 +77,43 @@ class FileWork:
             return paths
         except Exception as ex:
             logging.error(f"Incorrect path - {ex}")
+ 
+    def serialize_private_key(self, private_key: rsa.RSAPrivateKey) -> None:
+        """
+        Serialize the private key and save it to a file.
+
+        :param private_key: The private key.
+        """
+        try:
+            with open(self.path, 'wb') as key_file:
+                key_file.write(private_key.private_bytes(encoding=serialization.Encoding.PEM,
+                                                         format=serialization.PrivateFormat.TraditionalOpenSSL,
+                                                         encryption_algorithm=serialization.NoEncryption()))
+        except Exception as e:
+            logging.error(f"Error in serializing private key - {e}")
+
+    def serialize_public_key(self, public_key: rsa.RSAPublicKey) -> None:
+        """
+        Serialize the public key and save it to a file.
+
+        :param public_key: The public key.
+        """
+        try:
+            with open(self.path, 'wb') as key_file:
+                key_file.write(public_key.public_bytes(encoding=serialization.Encoding.PEM,
+                                                       format=serialization.PublicFormat.SubjectPublicKeyInfo))
+        except Exception as e:
+            logging.error(f"Error in serializing public key - {e}")
+
+    def deserialize_private_key(self) -> rsa.RSAPrivateKey:
+        """
+        Deserialize the private key and return it.
+
+        :return: The deserialized private key.
+        """
+        try:
+            with open(self.path, 'rb') as key_file:
+                return serialization.load_pem_private_key(key_file.read(), password=None)
+        except Exception as e:
+            logging.error(f"Error in deserializing private key - {e}")
